reduce warnings 2

+++ nclextra/msetsblock.cpp
@@ -312,7 +312,7 @@ CharSet *MSetsBlock::findCharSet(string name) {
 }
 
 int MSetsBlock::findArea(string &name) {
-	for (int i = 0; i < sets.size(); i++)
+	for (size_t i = 0; i < sets.size(); i++)
 		if (sets[i]->name == name) return i;
 	return -1;
 }
